==> notebooks/02_divstats.ipynb <==
Sys.setlocale(category = "LC_ALL", locale = "en_US.UTF-8")

# Dependencies

# Functions

# compute_divstats_parallel <- function(lst, chain, batchname, outpath, cores = max(1, parallel::detectCores() - 1)) {
#     if (!requireNamespace("iNEXT", quietly = TRUE)) stop("Package 'iNEXT' is required.")
#     if (!requireNamespace("parallel", quietly = TRUE)) stop("Package 'parallel' is required.")

#     sample_names <- names(lst)
#     if (is.null(sample_names) || any(!nzchar(sample_names))) {
#         sample_names <- paste0("sample_", seq_along(lst))
#     }

#     iNEXT_list <- parallel::mclapply(
#         X = lst,
#         FUN = function(x) iNEXT::iNEXT(x, q = 0, datatype = "abundance"),
#         mc.cores = cores
#     )
#     names(iNEXT_list) <- sample_names

#     extract_qD_max <- function(res) {
#         est <- res$iNextEst
#         if (is.list(est)) est <- est[[1]]
#         if (!is.data.frame(est) || !"qD" %in% names(est)) return(NA_real_)
#         max(est$qD, na.rm = TRUE)
#     }
#     extract_SC_max <- function(res) {
#         est <- res$iNextEst
#         if (is.list(est)) est <- est[[1]]
#         if (!is.data.frame(est) || !"SC" %in% names(est)) return(NA_real_)
#         max(est$SC, na.rm = TRUE)
#     }
#     extract_asy <- function(res, diversity_label) {
#         ae <- res$AsyEst
#         if (!is.data.frame(ae)) return(c(NA_real_, NA_real_, NA_real_))
#         row <- ae[ rownames(ae) == diversity_label, c("Observed", "Estimator", "Est_s.e."), drop = FALSE]
#         print(row)
#         if (nrow(row) == 0) return(c(NA_real_, NA_real_, NA_real_))
#         as.numeric(row[1, ])
#     }

#     chain_len  <- sapply(lst, length)
#     reads_sum  <- sapply(lst, sum)
#     cpkr       <- ifelse(reads_sum > 0, (chain_len / reads_sum) * 1000, NA_real_)
#     avg_reads  <- sapply(lst, mean)
#     vmr        <- sapply(lst, function(x) { m <- mean(x); if (m > 0) stats::sd(x) / m else NA_real_ })
#     max_reads  <- sapply(lst, max)
#     singletons <- sapply(lst, function(x) sum(x == 1))
#     doubletons <- sapply(lst, function(x) sum(x == 2))

#     qD_max <- sapply(iNEXT_list, extract_qD_max)
#     SC_max <- sapply(iNEXT_list, extract_SC_max)

#     richness <- do.call(rbind, lapply(iNEXT_list, extract_asy, diversity_label = "Species Richness"))
#     shannon  <- do.call(rbind, lapply(iNEXT_list, extract_asy, diversity_label = "Shannon diversity"))
#     simpson  <- do.call(rbind, lapply(iNEXT_list, extract_asy, diversity_label = "Simpson diversity"))
# print(shannon)
#     div_stats <- matrix(ncol = 19, nrow = length(lst))
#     colnames(div_stats) <- c(
#         chain, "Reads", "CPKR", "Average_reads", "VMR", "Max_reads",
#         "Singletons", "Doubletons", "qD", "Sample_Coverage",
#         "observed_Richness", "estimated_Richness", "SE_Richeness",
#         "observed_Shannon", "estimated_Shannon", "SE_Shannon",
#         "observed_Simpson", "estimated_Simpson", "SE_Simpson"
#     )
#     rownames(div_stats) <- sample_names

#     div_stats[, chain] <- chain_len
#     div_stats[, "Reads"] <- reads_sum
#     div_stats[, "CPKR"] <- cpkr
#     div_stats[, "Average_reads"] <- avg_reads
#     div_stats[, "VMR"] <- vmr
#     div_stats[, "Max_reads"] <- max_reads
#     div_stats[, "Singletons"] <- singletons
#     div_stats[, "Doubletons"] <- doubletons
#     div_stats[, "qD"] <- qD_max
#     div_stats[, "Sample_Coverage"] <- SC_max
#     div_stats[, c("observed_Richness", "estimated_Richness", "SE_Richeness")] <- richness
#     div_stats[, c("observed_Shannon", "estimated_Shannon", "SE_Shannon")] <- shannon
#     div_stats[, c("observed_Simpson", "estimated_Simpson", "SE_Simpson")] <- simpson

#     out_file <- file.path(outpath, paste0("divstats_", chain, batchname, ".csv"))
#     utils::write.csv(div_stats, file = out_file, row.names = TRUE)

#     invisible(list(div_stats = div_stats, iNEXT_list = iNEXT_list))
# }

source("/Users/anabbi/git/ped_CapTCRseq/R/Immune_Diversity.R")

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Plots/"
manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Manifests/"

gitpath <- "/Users/anabbi/git/ped_CapTCRseq/"

# Main

h4hpath <- "/Users/anabbi/Desktop/H4H/INTERCEPT/data/mixcr/"

## TRB

file_list <- list.files(h4hpath, pattern = paste0("CLONES_", "TRB"), recursive = T)

tail(file_list)

Some files are duplicated. Make a table to identify and pick the larger clone file

file_df <- as.data.frame(file_list)

file_df$samplename <- gsub(".*CLONES_TRB", "", file_df$file_list)
file_df$samplename <- gsub(".txt", "", file_df$samplename)

file_df$size <- NA
for (i in 1:nrow(file_df)) {
    file_df$size[i] <- file.size(paste0(h4hpath, file_df$file_list[i]))
}


head(file_df)

file_df[ grepl("395-01", file_df$file_list),]

#if duplicated keep the bigger file corresponding to deeper seqeuncing.
file_df_dedup <- file_df[0,]

for(s in unique(file_df$samplename)){
    tmp <- file_df[ file_df$samplename == s,]
    if(nrow(tmp) ==1 ){
        file_df_dedup <- rbind(file_df_dedup, tmp)
    }
    if(nrow(tmp) > 1 ){
        file_df_dedup <- rbind(file_df_dedup, tmp[ tmp$size == max(tmp$size),])
    }    
}

file_df_dedup[ grepl("343", file_df_dedup$file_list),]

file_df_dedup[ grepl("343-05", file_df_dedup$file_list),]

dim(file_df)
dim(file_df_dedup)

file_df_dedup_trb <- file_df_dedup
# save(file_df_dedup_trb, file = paste0(datapath,"capTCRseq/TRBfiles.RData"))

batchTRB <- immunelistfx(file_df_dedup$file_list, h4hpath, "TRB")

names(batchTRB) <- gsub(".*CLONES_TRB", "", names(batchTRB))

head(names(batchTRB))

TRBlist <- batchTRB 

readr::write_rds(TRBlist, file = paste0(gitpath,"Data/TRBlist.rds"))

save(TRBlist, file = paste0(datapath,"capTCRseq/TRBlist.RData"))

### divstats

load(file = paste0(datapath,"capTCRseq/TRBlist.RData"))
load(file = paste0(datapath,"capTCRseq/TRBfiles.RData"))

TRBlist <- TRBlist[sapply(TRBlist, function(x) length((x)) != 0)]
length(names(TRBlist))

# add filenames to divstats
myfilenames <- file_df_dedup_trb[ file_df_dedup_trb$samplename %in% names(TRBlist),]

head(myfilenames)

save(TRBlist, file = "/Users/anabbi/Desktop/sam/intercept/TRBlist.RData")
save(file_df_dedup_trb, file = "/Users/anabbi/Desktop/sam/intercept/TRBfiles.RData")

# parallel_inext.R ## run on valinor interactive 

### Cleanup divstats

divstats <- read.csv(paste0(datapath, "capTCRseq/divstats_TRBCHP_updated.csv"), header = T, stringsAsFactors = F, row.names = 1)

# add 9 samples with nonproductive cdr3
addsamples <- c(
    "CHP_347-01-cfDNA", "CHP_366-05_cfDNA", "CHP_377-01-cfDNA",
    "CHP_381-05_cfDNA", "CHP_391-02_cfDNA", "CHP_392-01_cfDNA",
    "CHP_398-01_cfDNA", "CHP_401-03_cfDNA", "CHP_419-FFPE-DNA_S2"
)


divstats[ nrow(divstats) + 9 , ] <- NA

rownames(divstats)[is.na(divstats$TRB)] <- addsamples

#example sample_id CHP_xxx_xx
divstats$sample_id <- rownames(divstats)

# remove trailing _cfDNA or PBMC-DNA from sample_id
divstats$sample_id <- gsub("_cfDNA$|_PBMC_DNA$|-cfDNA$|-PBMC_DNA$|-PBMC-DNA$|_PBMC-DNA", "", divstats$sample_id)


#example basename CHP_xxx-xx-PBMC-DNA
divstats$basename <- rownames(divstats)

# add filenames to divstats
divstats$file <- myfilenames$file_list[ match(divstats$basename, myfilenames$samplename)]

head(divstats[grepl("CHP_407", divstats$sample_id),])

head(divstats[grepl("CHP_407", divstats$sample_id),])

head(myfilenames[grepl("CHP_407", myfilenames$samplename),])

# Fix sample_id for CHP_407-03__PBMC-DN (double underscore issue)
divstats$sample_id <- gsub("__PBMC-DN$", "", divstats$sample_id)
divstats$sample_id <- gsub("_-PBMC-DN$", "", divstats$sample_id)


divstats$sample_id <- gsub("[-_]$", "", divstats$sample_id)

Separate samples

#divstats$Sample <- NULL

divstats$Sample <- NA
divstats$Sample[grepl("PBMC", rownames(divstats))] <- "PBMC"
divstats$Sample[grepl("cfDNA", rownames(divstats))] <- "cfDNA"
divstats$Sample[grepl("FFPE", rownames(divstats))] <- "Tumor"

Separate LFS and non-LFS

divstats$Group <- NA
divstats$Group[grepl("LFS", rownames(divstats))] <- "LFS"
divstats$Group[grepl("CHP", rownames(divstats))] <- "CHP"

summary(divstats$Reads)

Get patient ids

divstats$Patient <- NA
# CHP
divstats$Patient[divstats$Group == "CHP"] <- substr(rownames(divstats)[divstats$Group == "CHP"],1,7)
# Normal samples have either YO or MO
divstats$Patient[grepl("YO", rownames(divstats))] <- gsub("_PBMC_DNA", "", rownames(divstats)[grepl("YO", rownames(divstats))])
divstats$Patient[grepl("MO", rownames(divstats))] <- gsub("_PBMC_DNA", "", rownames(divstats)[grepl("MO", rownames(divstats))])
# LFS
divstats$Patient[divstats$Group == "LFS"] <- substr(rownames(divstats)[divstats$Group == "LFS"],1,12)

Get cycles. For LFS, Normals and Tumors recode to 00

divstats$Cycle <- NA
divstats$Cycle[divstats$Group == "CHP"] <- substr(rownames(divstats)[divstats$Group == "CHP"],9,10)
divstats$Cycle[grepl("YO", rownames(divstats))] <- "00"
divstats$Cycle[grepl("MO", rownames(divstats))] <- "00"
divstats$Cycle[ divstats$Sample == "Tumor"] <- "00"
divstats$Cycle[ divstats$Group == "LFS"] <- "00"

divstats$sample_id <- NA

# bind patient and cycle for PBMC and cfdna to generate sample id
divstats$sample_id[divstats$Sample != "Tumor"] <- paste(divstats$Patient[divstats$Sample != "Tumor"], "_",
    divstats$Cycle[divstats$Sample != "Tumor"],
    sep = ""
)

# For tumor samples start with rownames because some of them are two biopsies from the same tumor
divstats$sample_id[divstats$Sample == "Tumor"] <- gsub("-FFPE-DN.*", "", rownames(divstats)[divstats$Sample == "Tumor"])


divstats$sample_id <- gsub("-", "_", divstats$sample_id)

divstats$Patient <- NULL
divstats$Cycle <- NULL
divstats$Sample <- NULL
divstats$Group <- NULL

write.csv(divstats,
    paste0(datapath, "capTCRseq/divstats_TRBCHP_updated_corrected.csv"),
    row.names = TRUE
)
