==> notebooks/03_TCR_metadata.ipynb <==
Sys.setlocale(category = "LC_ALL", locale = "en_US.UTF-8")

# Dependencies

pacman::p_load(ggplot2)

# Functions

source("/Users/anabbi/git/ped_CapTCRseq/R/ggplot2_theme.R")
source("/Users/anabbi/git/ped_CapTCRseq/R/color_schemes.R")
source("/Users/anabbi/git/ped_CapTCRseq/R/Misc_functions.R")

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Plots/"
manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Manifests/"
gitpath <- "/Users/anabbi/git/ped_CapTCRseq/"

# Main

metadata <- read.csv(
    file = paste0(datapath, "INT_metadata_flow.csv"),
    header = T, stringsAsFactors = F, colClasses = c("cycle" = "character"), check.names = F
)


head(metadata)

colnames(metadata)

# confirmed with David, remove those samples with NCR% > 100
metadata$`TIM3%`[metadata$`TIM3%` > 100] <- NA
metadata$`LAG3%`[metadata$`LAG3%` > 100] <- NA
metadata$`PD1%`[metadata$`PD1%` > 100] <- NA
metadata$`PDL1%`[metadata$`PDL1%` > 100] <- NA

## TRB

divstats <- read.csv(paste0(datapath, "capTCRseq/divstats_TRBCHP_updated_corrected.csv"),
    header = T, stringsAsFactors = F, row.names = 1
)

head(divstats[ divstats$sample_id == "CHP_407_04",])

divstats$sampletype <- NA
divstats$sampletype[grepl("PBMC", divstats$basename)] <- "PBMC"
divstats$sampletype[grepl("cfDNA", divstats$basename)] <- "cfDNA"
divstats$sampletype[grepl("FFPE", divstats$basename)] <- "FFPE"

table(divstats$sampletype, useNA = "always")

divstats[grepl("CHP_323", rownames(divstats)),]

lfs <- divstats$sample_id[grepl("LFS", divstats$sample_id)]
normals <- divstats$sample_id[grepl("YO", divstats$sample_id) | grepl("MO", divstats$sample_id)]

table(divstats$sampletype, useNA = "always")

#merge div file separately for each sample type, then merge, just to capture all data regardless of available datatype
meta_div_cf <- merge(metadata, divstats[divstats$sampletype == "cfDNA",], by = "sample_id", all = FALSE)
meta_div_pbmc <- merge(metadata, divstats[divstats$sampletype == "PBMC",], by = "sample_id", all = TRUE)

dim(metadata)
dim(meta_div_pbmc)
dim(divstats[divstats$sampletype == "PBMC",])

meta_div_cf[ meta_div_cf$sample_id == "CHP_407_01",]

div_ffpe <- divstats[divstats$sampletype == "FFPE",]

div_ffpe$Patient <- gsub("_A", "", div_ffpe$sample_id)
div_ffpe$Patient <- gsub("_B", "", div_ffpe$Patient)

meta_div_ffpe <- cbind(metadata[ match(div_ffpe$Patient, metadata$Patient),], div_ffpe)

which(colnames(meta_div_ffpe) == "sample_id")

meta_div_ffpe <- meta_div_ffpe[, c(36,1:35,38:ncol(meta_div_ffpe)-1)]

meta_div_ffpe$sample_id.1 <- NULL

head(meta_div_ffpe)

meta_div <- rbind(meta_div_cf, meta_div_pbmc, meta_div_ffpe)

dim(meta_div)

#create a group for LFS and non-LFS
meta_div$Group <- NA
meta_div$Group[grepl("LFS", meta_div$sample_id)] <- "LFS"
meta_div$Group[grepl("CHP", meta_div$sample_id)] <- "CHP"

#create sample type
meta_div$sampletype <- NA
meta_div$sampletype[grepl("PBMC", meta_div$basename)] <- "PBMC"
meta_div$sampletype[grepl("cfDNA", meta_div$basename)] <- "cfDNA"
meta_div$sampletype[grepl("FFPE", meta_div$basename)] <- "Tumor"
#sampletyoe is PBMC if basename is NA, they are flow data
meta_div$sampletype[is.na(meta_div$basename)] <- "PBMC"

# add Patient id for normals, lfs
meta_div$Patient[meta_div$sample_id %in% c(lfs, normals)] <-
       substr(
              meta_div$sample_id[meta_div$sample_id %in% c(lfs, normals)], 1,
              nchar(meta_div$sample_id[meta_div$sample_id %in% c(lfs, normals)]) - 3
       )


# few samples we sequenced but was not in David's file
meta_div$Patient[is.na(meta_div$Patient) & meta_div$sampletype != "Tumor"] <-
       substr(
              meta_div$sample_id[is.na(meta_div$Patient) & meta_div$sampletype != "Tumor"], 1,
              nchar(meta_div$sample_id[is.na(meta_div$Patient) & meta_div$sampletype != "Tumor"]) - 3
       )


# Patient id same as sampel_id for tumor samples
meta_div$Patient[is.na(meta_div$Patient)] <- meta_div$sample_id[is.na(meta_div$Patient)]


# overwrite cycle for consistency

meta_div$cycle <- NA
meta_div$cycle <- substr(
       meta_div$sample_id,
       nchar(meta_div$sample_id) - 1, nchar(meta_div$sample_id)
)


#add X to prevent numeric
meta_div$cycle <- paste0("X", meta_div$cycle)

# cycle for tumors are 00
meta_div$cycle[ which(meta_div$sampletype == "Tumor")] <- "X00"

table(meta_div$cycle)

#Disease_type for normals = Normal and lfs = LFS
meta_div$Disease_type[ meta_div$sample_id %in% normals] <- "Normal"

meta_div$Disease_type[ meta_div$sample_id %in% lfs] <- "LFS"

table(meta_div$Disease_type, useNA = "always")

#minor fix
meta_div$Patient[ meta_div$Patient == "CHP_404_A"] <- "CHP_404"
meta_div$Patient[ meta_div$Patient == "CHP_404_B"] <- "CHP_404"
meta_div$Patient[ meta_div$Patient == "CHP_344_A"] <- "CHP_344"

unique(meta_div$Patient)

# add dx for tumors
for (i in unique(meta_div$Patient[which(meta_div$sampletype == "Tumor")])) {
   # print(i)
   tmp <- meta_div[meta_div$Patient == i, ]
   cancer <- names(table(tmp$Disease_type)[1])
   meta_div$Disease_type[meta_div$Patient == i &
      meta_div$sampletype %in% "Tumor"] <- cancer
}


meta_div[ is.na(meta_div$Disease_type),]

# old samples. Remove them!
meta_div <- meta_div[ !is.na(meta_div$Disease_type),]

table(meta_div$cancergroup, useNA = "always")

table(meta_div$Disease_type, useNA = "always")

meta_div$cancergroup[ meta_div$sampletype == "Tumor"] <- "Solid"
meta_div$cancergroup[ meta_div$Disease_type == "Normal"] <- "Normal"
meta_div$cancergroup[ meta_div$Disease_type == "LFS"] <- "LFS"

table(meta_div$cancergroup, useNA = "always")

table(meta_div$sampletype, useNA = "always")

head(meta_div[meta_div$sample_id == "CHP_407_01",])

length(unique(meta_div$Patient))

meta_div[ meta_div$Patient == "CHP_407",]

save(meta_div, file = paste0(datapath,"capTCRseq/meta_div_TRB.RData"))
