==> notebooks/07_Immunarch_data.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#Diversity-using-hill-numbers-from-immunarch" data-toc-modified-id="Diversity-using-hill-numbers-from-immunarch-4.1">Diversity using hill numbers from immunarch</a></span></li><li><span><a href="#Create-metadata-for-immunarch" data-toc-modified-id="Create-metadata-for-immunarch-4.2">Create metadata for immunarch</a></span></li></ul></li></ul></div>

# Dependencies

library(ggplot2)
library(immunarch)
library(ggbeeswarm)

# Functions

histp <- function(df, var){
    myp <- ggplot(data = df, aes(x = eval(parse(text = var)))) + 
    geom_histogram() + myplot + myaxis + labs(x = var)
    return(myp)
}

source("/Users/anabbi/git/ped_CapTCRseq/R/ggplot2_theme.R")
source("/Users/anabbi/git/ped_CapTCRseq/R/color_schemes.R")

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Plots/"
manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Manifests/"

gitpath <- "/Users/anabbi/git/ped_CapTCRseq/"

# Main

# Copied back from h4h
load(paste0(datapath, "capTCRseq/int_load.RData"))

#remove scores from V and J
for(i in 1:length(int_load$data)){
    int_load$data[[i]] <- as.data.frame(int_load$data[[i]])
    int_load$data[[i]][,"V.name"] <- gsub("\\*00\\(.*", "", int_load$data[[i]][,"V.name"])
    int_load$data[[i]][,"D.name"] <- gsub("\\*00\\(.*", "", int_load$data[[i]][,"D.name"])
    int_load$data[[i]][,"J.name"] <- gsub("\\*00\\(.*", "", int_load$data[[i]][,"J.name"])
    int_load$data[[i]] <- as_tibble(int_load$data[[i]])
}

# names cleanup
names(int_load$data) <- gsub("CLONES_TRB", "", names(int_load$data))

int_load$meta$Sample <- gsub("CLONES_TRB", "", int_load$meta$Sample)

head(names(int_load$data))

head(int_load$meta)

## Diversity using hill numbers from immunarch

load(file = paste0(datapath,"capTCRseq/meta_div_TRB.RData"))

div_hill_aa <- repDiversity(int_load$data, "hill", .min.q = 0)
div_hill_v <- repDiversity(int_load$data, "hill", .col = "v", .min.q = 0)
div_hill_j <- repDiversity(int_load$data, "hill", .col = "j", .min.q = 0)

div_hill_aav <- repDiversity(int_load$data, "hill", .col = "aa+v", .min.q = 0)

Add to divstats

q0_aa <- as.data.frame(div_hill_aa[ div_hill_aa$Q == 0,])
q0_v <- as.data.frame(div_hill_v[ div_hill_v$Q == 0,])
q0_j <- as.data.frame(div_hill_j[ div_hill_j$Q == 0,])
q0_aav <- as.data.frame(div_hill_aav[ div_hill_aav$Q == 0,])

head(q0_aa,15)

meta_div$q0_aa <- q0_aa$Value[match(meta_div$basename, q0_aa$Sample)]
meta_div$q0_v <- q0_v$Value[match(meta_div$basename, q0_v$Sample)]
meta_div$q0_j <- q0_j$Value[match(meta_div$basename, q0_j$Sample)]
meta_div$q0_aav <- q0_aav$Value[match(meta_div$basename, q0_aav$Sample)]
meta_div$q0_vj <- rowMeans(meta_div[,c("q0_j", "q0_v")])

q1_aa <- as.data.frame(div_hill_aa[ div_hill_aa$Q == 1,])
q1_v <- as.data.frame(div_hill_v[ div_hill_v$Q == 1,])
q1_j <- as.data.frame(div_hill_j[ div_hill_j$Q == 1,])
q1_aav <- as.data.frame(div_hill_aav[ div_hill_aav$Q == 1,])

meta_div$q1_aa <- q1_aa$Value[match(meta_div$basename, q1_aa$Sample)]
meta_div$q1_v <- q1_v$Value[match(meta_div$basename, q1_v$Sample)]
meta_div$q1_j <- q1_j$Value[match(meta_div$basename, q1_j$Sample)]
meta_div$q1_aav <- q1_aav$Value[match(meta_div$basename, q1_aav$Sample)]
meta_div$q1_vj <- rowMeans(meta_div[,c("q1_j", "q1_v")])

save(meta_div, file = paste0(datapath,"capTCRseq/divstats_immunarch.RData" ))

## Create metadata for immunarch

load(file = paste0(datapath,"capTCRseq/divstats_immunarch.RData" ))

metatmp <- merge(int_load$meta, meta_div, by.x = "Sample", by.y = "basename")

meta_div$

table(meta_div$cancergroup, useNA = "always")

table(metatmp$cancergroup, useNA = "always")

meta_div[ !meta_div$sample_id %in% metatmp$sample_id,]

#256 samples have flow data but either were not sequenced or didnt recover any productive cdr3

int_load$meta <- as_tibble(metatmp)

dim(int_load$meta)

length(int_load$data)

names(int_load$data)[! names(int_load$data) %in% int_load$meta$Sample]

#remove those old samples (patient numbers < 320)
rmsmpl <- c('CHP_303-05-PBMC-DNA', 'CHP_315-05_cfDNA', 'CHP_315-05-PBMC-DNA', 'CHP_319-05-PBMC-DNA')

int_load$data <- int_load$data[ !names(int_load$data) %in% rmsmpl]

length(int_load$data)

int_load$data <- int_load$data[ int_load$meta$Sample]

save(int_load, file = paste0(datapath, "capTCRseq/int_load_meta.RData"))
