==> notebooks/07_master_data_goodSamples.ipynb <==
Sys.setlocale(category = "LC_ALL", locale = "en_US.UTF-8")

# Dependencies

# Functions

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Plots/"
manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Manifests/"

gitpath <- "/Users/anabbi/git/ped_CapTCRseq/"

# Main

# PBMC

load(file = paste0(datapath, "capTCRseq/PBMC_GoodSamples.RData"))

head(subset_pbmc)

load(file = paste0(datapath,"capTCRseq/meta_div_TRB.RData"))

(meta_div[ meta_div$Patient == "CHP_407",])

dim(subset_pbmc)

length(meta_div$sampletype)

load(file = paste0(datapath, "capTCRseq/cfDNA_GoodSamples.RData"))

head(subset_cfdna)

table(meta_div$Disease_type)

meta_div_goodsamples <- meta_div[ which(meta_div$sampletype == "Tumor"),] # all tumours
meta_div_goodsamples <- rbind(
meta_div[ which(meta_div$basename %in% subset_pbmc$basename),], #good pbmcs
meta_div[ which(meta_div$basename %in% subset_cfdna$basename),], # good cfDNAs
meta_div[ is.na(meta_div$basename),], # all flow-only data
meta_div[ which(meta_div$Disease_type == "LFS"),], # add LFS
meta_div_goodsamples)


all407 <- (meta_div[ meta_div$Patient == "CHP_407",])
good407 <- (meta_div_goodsamples[ meta_div_goodsamples$Patient == "CHP_407",])

colnames(all407)

table(meta_div_goodsamples$sampletype, useNA = "ifany")

# a few PBMC samples failed QC for TCR, but Flow is still good. add those back in but remove TCR data
allpbmcsamples <- meta_div$sample_id[ meta_div$sampletype == "PBMC" & !is.na(meta_div$basename)]
goodpbmcsamples <- meta_div_goodsamples$sample_id[ meta_div_goodsamples$sampletype == "PBMC" ]

length(allpbmcsamples)
length(goodpbmcsamples)

goodflowpbmcsamples <- allpbmcsamples[! allpbmcsamples %in% goodpbmcsamples]
goodflowpbmcsamples


goodflow <- meta_div[ meta_div$sampletype == "PBMC",]
goodflow <- goodflow[ goodflow$sample_id %in% goodflowpbmcsamples,]

# Replace TCR-related columns with NA for flow-only PBMC samples
tcr_columns <- c('TRB', 'Reads', 'CPKR', 'Average_reads', 'VMR', 'Max_reads', 
                 'Singletons', 'Doubletons', 'qD', 'Sample_Coverage', 
                 'observed_Richness', 'estimated_Richness', 'SE_Richeness', 
                 'observed_Shannon', 'estimated_Shannon', 'SE_Shannon', 
                 'observed_Simpson', 'estimated_Simpson', 'SE_Simpson', 
                 'basename', 'file')

goodflow[, tcr_columns] <- NA

colnames(goodflow)


meta_div_goodsamples <- rbind(meta_div_goodsamples, goodflow)

length(unique(meta_div_goodsamples$Patient[ meta_div_goodsamples$sampletype == "PBMC"]))

meta_div_goodsamples[ meta_div_goodsamples$sample_id == "CHP_407_03",]

table(meta_div_goodsamples$sampletype, useNA = "ifany")

meta_div_goodsamples$cancergroup[ which(meta_div_goodsamples$cancergroup == "Solid")] <- "Solid tumors" # rename solid tumors

meta_div_goodsamples$cancergroup[ meta_div_goodsamples$Disease_type == "T-ALL" ] <- "T-cell malignancies"
meta_div_goodsamples$cancergroup[ meta_div_goodsamples$Disease_type == "ALCL" ] <- "T-cell malignancies"

meta_div_goodsamples$log10shann <- log10(meta_div_goodsamples$observed_Shannon)

# recode cd3 = 0 to NA
meta_div_goodsamples$CD3[which(meta_div_goodsamples$CD3 == 0)] <- NA

head(meta_div_goodsamples)

meta_div_goodsamples$sample_id_sampletype <- paste0(meta_div_goodsamples$sample_id, "_", meta_div_goodsamples$sampletype)

# remove samples with CD3 == 0, TRB == NA and sampletype == PBMC
# only clinical + ex vivo data
rmsamples <- meta_div_goodsamples$sample_id_sampletype[ is.na(meta_div_goodsamples$CD3) & is.na(meta_div_goodsamples$TRB) & meta_div_goodsamples$sampletype == "PBMC"]

table(meta_div_goodsamples$sampletype, useNA = "ifany")

dim(meta_div_goodsamples)

meta_div_goodsamples <- meta_div_goodsamples[ !meta_div_goodsamples$sample_id_sampletype %in% rmsamples,]

dim(meta_div_goodsamples)

table(meta_div_goodsamples$sampletype, useNA = "ifany") # 246 samples removed they had only clinical and ex vivo data

meta_div_goodsamples[ meta_div_goodsamples$sample_id == "CHP_330_01",]

meta_div_goodsamples$assay <- NA
for (i in 1:nrow(meta_div_goodsamples)) {
    if (is.na(meta_div_goodsamples$CD3[i]) & is.na(meta_div_goodsamples$TRB[i])) {
        meta_div_goodsamples$assay[i] <-  NA
    } else if (!is.na(meta_div_goodsamples$CD3[i]) & is.na(meta_div_goodsamples$TRB[i])) {
        meta_div_goodsamples$assay[i] <- "Flow cytometry"
    } else if (is.na(meta_div_goodsamples$CD3[i]) & !is.na(meta_div_goodsamples$TRB[i])) {
        meta_div_goodsamples$assay[i] <- "capTCR-seq"
    } else if (!is.na(meta_div_goodsamples$CD3[i]) & !is.na(meta_div_goodsamples$TRB[i])) {
        meta_div_goodsamples$assay[i] <- "Flow cytometry & capTCR-seq"
    }
}

meta_div_goodsamples$assay[ meta_div_goodsamples$sampletype == "cfDNA"] <- "capTCR-seq"


table(meta_div_goodsamples$sampletype)

table(meta_div_goodsamples$assay, useNA = "ifany")

meta_div_goodsamples[ is.na(meta_div_goodsamples$assay),]

meta_div_goodsamples <- meta_div_goodsamples[ !is.na(meta_div_goodsamples$assay),] # 1 tumor sample removed

rmsamples <- meta_div_goodsamples$basename[ meta_div_goodsamples$assay == "capTCR-seq" & is.na(meta_div_goodsamples$TRB)] # if sequenced but no clones

dim(meta_div_goodsamples)

meta_div_goodsamples <- meta_div_goodsamples[ !(meta_div_goodsamples$basename %in% rmsamples),]

dim(meta_div_goodsamples)

dim(meta_div_goodsamples)

table(meta_div_goodsamples$sampletype)

table(meta_div_goodsamples$assay)

# remove samples not discussed in the paper
#remove Tumor samples
meta_div_goodsamples <- meta_div_goodsamples[ which(meta_div_goodsamples$sampletype != "Tumor"),]
#remove Normal samples
meta_div_goodsamples <- meta_div_goodsamples[ which(meta_div_goodsamples$cancergroup != "Normal"),]
# remove LFS samples
meta_div_goodsamples <- meta_div_goodsamples[which(meta_div_goodsamples$cancergroup != "LFS"),]

table(meta_div_goodsamples$sampletype)

readr::write_rds(meta_div_goodsamples, paste0(datapath, "capTCRseq/meta_div_goodsamples.rds"))
# push to git
readr::write_rds(meta_div_goodsamples, paste0(gitpath, "data/metadata_goodsamples.rds"))

length(unique(meta_div_goodsamples$Patient))

meta_div_goodsamples[ meta_div_goodsamples$Patient == "CHP_330",]

# subset_meta <- meta_div_goodsamples[, c("Patient", "sample_id","cycle", "Disease_type", "cancergroup", "sampletype", "CD3", "TRB", "assay")]

# table(subset_meta$sampletype)

# table(subset_meta$cancergroup)

# readr::write_csv(subset_meta, paste0(datapath, "metadata_schematic.csv"))

# SOC df

meta_div <- readr::read_rds(file = paste0(datapath,"capTCRseq/meta_div_goodsamples.rds"))

soc <- readxl::read_xlsx(paste0(manifestpath, "SOC_chemo_cycles_AN.xlsx"))

soc$Patient <- paste0("CHP_", soc$`Patient ID`)

soc_good <- soc[ soc$Patient %in% meta_div$Patient,]

soc_good <- soc_good[ , 1:which(grepl("Cycle 5", colnames(soc_good)))[4]]

soc_good$Disease <-soc_good$Diagnosis
soc_good$Disease[ soc_good$Disease == "Hepatoblastoma"] <- "HB"
soc_good$Disease[soc_good$Disease == "Ewings"] <- "EWS"
soc_good$Disease[soc_good$Disease == "Burkitts"] <- "BL"
soc_good$Disease[soc_good$Disease == "Burkitts "] <- "BL"
soc_good$Disease[soc_good$Disease == "Burkitt's"] <- "BL"
soc_good$Disease[soc_good$Disease == "Wilms"] <- "WILMS"
soc_good$Disease[soc_good$Disease == "B cell lymphoma"] <- "BLL"
soc_good$Disease[soc_good$Disease == "B lymphoblastic lymphoma"] <- "BLL"
soc_good$Disease[soc_good$Disease == "NBL"] <- "NB"
soc_good$Disease[soc_good$Disease == "aLL"] <- "ALL"

soc_good$Cancergroup <- meta_div$cancergroup[match(soc_good$Disease, meta_div$Disease_type)]

length(unique(meta_div$Patient))

dim(soc_good)

# Save the cleaned soc_good dataframe as CSV
write.csv(soc_good, file = "~/git/ped_CapTCRseq/chemo_cycles_cleaned.csv", row.names = FALSE)

# Save the cleaned soc_good dataframe as RDS
saveRDS(soc_good, file = file.path(datapath, "chemo_cycles_cleaned.rds"))
