==> notebooks/09_discovery_validation_sets.ipynb <==
# Dependencies

pacman::p_load(dplyr)

# Functions

source("/Users/anabbi/git/ped_CapTCRseq/R/ggplot2_theme.R")
source("/Users/anabbi/git/ped_CapTCRseq/R/color_schemes.R")

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Plots/"
manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/INTERCEPT/Manifests/"
gitpath <- "/Users/anabbi/git/ped_CapTCRseq/"

h4hpath <- "/Users/anabbi/Desktop/H4H/INTERCEPT/"

# Main

bind all metadata

load(file = paste0(datapath,"capTCRseq/meta_div_TRB.RData"))

load(file = paste0(datapath, "Adaptivedatasets/allAdaptive.RData"))

# remove those with no TCR data
meta_div <- meta_div[ !is.na(meta_div$file),]

meta_div$study <- "INTERCEPT"
meta_div$Agegroup <- "Child"
meta_div$Sex <- NA
meta_div$disease <- meta_div$cancergroup
meta_div$sample_name <- meta_div$basename

allAdaptive$sampletype <- "PBMC"
allAdaptive$sampletype[grepl("Cord", allAdaptive$sample_name )] <- "Cord blood"

# replace age with Age in allAdative
colnames(allAdaptive)[colnames(allAdaptive) == "age"] <- "Age"
# subject in allAdaptive is the same as Patient in meta_div
colnames(allAdaptive)[colnames(allAdaptive) == "subject"] <- "Patient"

mycols <- intersect(colnames(meta_div), colnames(allAdaptive))

mycols

all_metadata <- rbind(allAdaptive[,mycols], meta_div[,mycols])

dim(all_metadata)

head(all_metadata)

table(all_metadata$study)

#save all_metadata
save(all_metadata, file = paste0(datapath, "all_metadata.RData"))

## All PBMC data

# remove cfDNA samples
all_metadata_pbmc <- all_metadata[!grepl("cfDNA", all_metadata$sample_name),]

all_metadata_pbmc <- all_metadata_pbmc[!grepl("FFPE", all_metadata_pbmc$sample_name),]

dim(all_metadata_pbmc)

table(all_metadata_pbmc$study)

## Discovery set

discovery_metadata <- all_metadata_pbmc[ all_metadata_pbmc$study != "Mitchell_Denver",]

discovery_metadata <- discovery_metadata[ !grepl("Keck", discovery_metadata$sample_name),] # emerson validation set

#remove carey Cord samples
discovery_metadata <- discovery_metadata[ !grepl("Cord", discovery_metadata$sample_name),]

dim(discovery_metadata[ discovery_metadata$study == "INTERCEPT",])

dim(discovery_metadata)

#### sample intercept study

# subset to INTERCEPT study
tmp <- discovery_metadata[ discovery_metadata$study == "INTERCEPT",]

length(table(tmp$Patient)) # number of patients

myfreq <- as.data.frame(table(tmp$Patient, tmp$disease), stringsAsFactors = F)

#remove zero
myfreq <- myfreq[myfreq$Freq > 0,]

head(myfreq)

# randomly select 50% of patients from each disease group regardless of how many samples each patient has
set.seed(7)
myfreq %>% group_by(Var2) %>% sample_frac(0.5) -> discovery_sample_freq

discovery_sample_freq

# all patients in discovery set
int_discovery_patients <- discovery_sample_freq$Var1

length(int_discovery_patients)

dim(discovery_metadata)

dim(discovery_metadata[ discovery_metadata$study == "INTERCEPT",])

all_discovery_patients <- c(discovery_metadata$Patient[ discovery_metadata$study != "INTERCEPT"], int_discovery_patients)

discovery_metadata <- discovery_metadata[ discovery_metadata$Patient %in% all_discovery_patients,]

length(table(discovery_metadata$Patient[ discovery_metadata$study == "INTERCEPT"] ))

# CHP_407 is dupliucate sample one from lymphoma one from ffpe
dim(discovery_metadata)

table(discovery_metadata$study)

## Validation set

head(all_metadata_pbmc)

validation_metadata <- all_metadata_pbmc[ !all_metadata_pbmc$Patient %in% all_discovery_patients,]

dim(validation_metadata)

dim(discovery_metadata)

# save discovery and validation metadata
save(discovery_metadata, file = paste0(datapath, "discovery_metadata.RData"))
save(validation_metadata, file = paste0(datapath, "validation_metadata.RData"))
